Update on future timing
Changed call to future building right outside the gridview constructor.
Also changed the future to wait to be completed before updating the ability array. This is to iprove performance and to see if the page still jiggles when loading.

Still uncertain what the best practice for this type of data loading is.